+++ include/nn/layers_impl/batchnorm_layer.hpp
@@ -38,20 +38,20 @@ private:
   Tensor<T> running_var_;
 
   // Per-micro-batch state for training
-  std::unordered_map<int, Tensor<T>> micro_batch_inputs_;
-  std::unordered_map<int, Tensor<T>> micro_batch_normalized_;
-  std::unordered_map<int, Tensor<T>> micro_batch_mean_;
-  std::unordered_map<int, Tensor<T>> micro_batch_var_;
-  std::unordered_map<int, Tensor<T>> micro_batch_std_;
+  std::unordered_map<size_t, Tensor<T>> micro_batch_inputs_;
+  std::unordered_map<size_t, Tensor<T>> micro_batch_normalized_;
+  std::unordered_map<size_t, Tensor<T>> micro_batch_mean_;
+  std::unordered_map<size_t, Tensor<T>> micro_batch_var_;
+  std::unordered_map<size_t, Tensor<T>> micro_batch_std_;
 
 public:
   explicit BatchNormLayer(size_t num_features, T epsilon = T(1e-5),
                           T momentum = T(0.1), bool affine = true,
                           const std::string &name = "batchnorm");
 
-  Tensor<T> forward(const Tensor<T> &input, int micro_batch_id = 0) override;
+  Tensor<T> forward(const Tensor<T> &input, size_t micro_batch_id = 0) override;
   Tensor<T> backward(const Tensor<T> &grad_output,
-                     int micro_batch_id = 0) override;
+                     size_t micro_batch_id = 0) override;
 
   std::string type() const override;
   LayerConfig get_config() const override;
